Refactor: Decompose TXDetailsScreen and update theme
- Decomposed `TXDetailsScreen` by extracting UI components into new features: `CardTextForTxDetailsFeature` and `AmlAndButtonGetAmlForTXDetailsFeature`.
- Replaced standard Material components with custom ones like `CustomScaffoldWallet`, `CustomTopAppBar`, and `CustomBottomCard`.
- Extended the color scheme in `Theme.kt` to include `greenColor` and `backgroundContainerButtonLight`.
- Reorganized file structure by moving AML and BottomSheet related features into dedicated subpackages.
- Moved `displayDateForAML.kt` and created `toAmlType.kt` within a new `utils/aml` package.
