+++ app/src/main/java/com/profpay/wallet/ui/components/custom/CustomBottomCard.kt
@@ -6,9 +6,7 @@ import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Card
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -35,8 +33,7 @@ fun CustomBottomCard(
             modifier =
                 modifierColumn
                     .fillMaxSize()
-                    .padding(bottom = bottomPadding.dp)
-                    .verticalScroll(rememberScrollState()),
+                    .padding(bottom = bottomPadding.dp),
             verticalArrangement = verticalArrangement,
             horizontalAlignment = horizontalAlignment,
         ) {
