+++ app/src/main/java/com/profpay/wallet/ui/feature/wallet/tx_details/aml/KnowAMLFeature.kt
@@ -1,4 +1,4 @@
-package com.profpay.wallet.ui.feature.wallet.tx_details
+package com.profpay.wallet.ui.feature.wallet.tx_details.aml
 
 import StackedSnakbarHostState
 import android.annotation.SuppressLint
@@ -37,6 +37,7 @@ import com.profpay.wallet.R
 import com.profpay.wallet.bridge.view_model.wallet.TXDetailsViewModel
 import com.profpay.wallet.data.database.entities.wallet.TransactionEntity
 import com.profpay.wallet.ui.app.theme.RedColor
+import com.profpay.wallet.utils.aml.displayDateForAML
 import com.profpay.wallet.ui.screens.wallet.AMLType
 import kotlinx.coroutines.launch
 import org.server.protobuf.aml.AmlProto
